api/requirements.txt
sqlalchemy==1.3.15
pymysql==0.9.3
flask-restful==0.3.8
flask-sqlalchemy==2.4.1
flask-marshmallow==0.11.0
marshmallow-sqlalchemy==0.22.3
Flask-Table==0.5.0
